Move types as-is to a separate crate.
