+++ crates/rsonpath-syntax/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "rsonpath-syntax"
+version = "0.1.0"
+authors = ["Mateusz Gienieczko <mat@gienieczko.com>"]
+description = "A JSONPath parser."
+readme = "README.md"
+license = "MIT"
+keywords = ["json", "jsonpath", "query", "search", "parser"]
+exclude = ["/tests"]
+categories = [ "parser-implementations", "text-processing"]
+repository = "https://github.com/V0ldek/rsonpath"
+workspace = "../.."
+edition = "2021"
+
+[package.metadata.docs.rs]
+rustdoc-args = ["--cfg", "docsrs"]
+all-features = true
+
+[dependencies]
+arbitrary = { version = "1.3.1", features = ["derive"], optional = true }
+nom = "7.1.3"
+thiserror = "1.0.50"
+
+[dev-dependencies]
+pretty_assertions = "1.4.0"
+test-case = "3.2.1"
+
+[features]
+default = []
+arbitrary = ["dep:arbitrary"]
