+++ crates/rsonpath-syntax/src/number.rs
