+++ crates/rsonpath-syntax/src/string.rs
