+++ fuzz/fuzz_targets/fuzz_arbitrary_json.rs
@@ -2,11 +2,11 @@
 
 use libfuzzer_sys::{arbitrary::Arbitrary, fuzz_target, Corpus};
 use rsonpath::input::BorrowedBytes;
-use rsonpath::query::JsonPathQuery;
 use rsonpath::{
     engine::{Compiler, Engine, RsonpathEngine},
     query::error::CompilerError,
 };
+use rsonpath_syntax::JsonPathQuery;
 use std::{
     collections::HashMap,
     fmt::{Debug, Display},
