+++ fuzz/fuzz_targets/query_fuzz_arbitrary_bytes.rs
@@ -1,7 +1,7 @@
 #![no_main]
 
 use libfuzzer_sys::fuzz_target;
-use rsonpath::query::JsonPathQuery;
+use rsonpath_syntax::JsonPathQuery;
 
 fuzz_target!(|data: &str| {
     let _ = JsonPathQuery::parse(data);
